Refactor for the integration forms/formsets
Moved the logic from the AccountFormSet class to the AccountForm
classes.  This allows for a greater variety of forms.  Before there
were a lot of requirements for the form classes, now basically
anything should work.  I'd like to add some forms for youtube accounts
and they weren't going to fit with the old system.

Started using the nose assertion functions.  I like they PEP-8 syntax
and not having to type "self.".