Move youtube account add/remove logic into forms
These forms are pretty trivial at this point, but this allows us to
add more complexity for youtube accounts.  I want to add a field to
update sync_teams and this seems like the best way to do it.

Added ability for account forms to redirect to other pages (needed for
the AddYouTubeAccountForm).  Updated AccountFormset a bit.

Added field to render a button.

Removed the views tests.  We were only testing the redirect logic and
that's now tested by the forms tests.