Added form field to set sync_teams
Updated template/CSS to display the widget.

Pass the user viewing the form to AccountFormset and
YoutubeAccountForm.  We need this to figure out which teams to allow
the admin to pick.

Added way to create a team and an admin member quickly in
utils.factories.

Removed unused youtube_accounts variable.