Fig/docker update
Fixed bin/dev to properly start a server.  Using fig run doesn't work,
because it doesn't expose the ports.  Changed "resetdb" to "dbreset".
Somehow this sounds better to me.  Made the bash command work with
older fig versions.

Fixed fig.yml to make the ssl-app and worker work correctly.  Changed
the entries for dev stuff so that the entrypoint is /usr/bin/env.  We
basically want to unset the entrypoint and I think this is a good hack
to simulate doing that.  Removed working_dir parts.  We are specifying
the same value is in our Dockerfile and working_dir is not supported
in fig 0.4.2.

Set the LANG env variable.  This prevents an exception in
createsuperuser.