Slight cleanup for the Workflow model
titleEdited was not really doing anything.  We were setting
titleEdited to true anytime we moved to the title stage and we were
only checking it to move past that stage.

showOverlay was also pretty weird.  We would set it to false, then set
it to true, then set it back to false immediately after.  Changed it
so that it starts true, then goes to false in appActionDone().