Editor HTML refactor
The HTML file was getting pretty huge.  Used django templating to
split it up.  This should also make it easier to refactor the layout

Tried to make the indentation consistent at 4 spaces per indent.  Now
that stuff is in different files we have more width to work with.

Lots of changed lines, but basically no changes to the HTML we render
other than whitespace changes.  The one change was to add a couple
missing </button> tags to properly close things.

Moved all the templates into the editor subdirectory.  I actually
think that we should have an editor app eventually and move the code
out of the subtitles app.