Removed collab stuff from the editor
We had a bunch of code here but little functionality.  The only thing
it provided was notes and task approval.  I'm going to be reworking
both of those as part of some other tickets, so I figure the simplest
thing to do is just rip it out and rebuild from scratch.