Reworked editor layout
Changed basically everything to use absolute positioning.  I think
this is the most logic way to do things because we want the content to
fit exactly on the page.  It also makes it simpler to move things
around and fixes the bug I was seeing when I changed my zoom size.