Don't update feeds more than once an hour
This can happen if our tasks get backed up in the queue and we have
multiple update_from_feed tasks queued up.  When we process each one
of those tasks, we create many other tasks, so it takes a really long
time to get through the entire queue.

The quick fix is to only queue up the other tasks if the feed hasn't
been recently updated.