Workflow button fixes (#1675)
Changed text back to "Accept" from "Approve".

Fixed color/icon for the send back button.

Changed button CSS so that we don't have a border, we just add an
extra px of padding.  This makes it so that we don't have to worry
about the border color when changing the background color.