Rework workflow content (#1675)
Add a linebreak before the edit title and description link

Added some extra text to explain steps.

Styled text inside steps better.

Make the content inside of the steps line up with the left side of the
heading, rather then the left side of the step number.

Display an error warning if you're on the review step, but don't have
completed subtitles.

Removed code to handle the tasks buttons.  We're going to do a more
dramatic change for approve/review tasks.