Re-implemented review/approve mode in the editor
In an attempt to try to clean up the handling of different modes, I
used to behaviors module to implement this.  It's a bit of overkill
right now, but I think it will help keep things sane when we add the
collab model modes to the list.

Still to be done is implementing the JS/HTML/CSS for the review mode.

Created documentation for the editor work modes and the behaviors
module.  Switched to using napoleon to handle this.  The docstrings
are so much nicer than the default sphinx docstrings.

Added a couple of convenience methods to Team and Task