Added API for subtitle actions
Started a new app for our API.  There are a bunch of issues with the
old one:
  - Uses a really old version of tastypie
  - Code is pretty confusing
  - Code is in our private repository, but only some of it needs to be

The new code feels way nicer to me, but we need to use some footwork
to to avoid package name conflicts.

I named the package "api", since "apiv2" is taken by the other app.
"api" used to be used for version 1 of the API, but we have not
supported that for a while now.

Made the new app use django-rest-framework.  It seems like a fine
package, but the main reason for the switch from tastypie is to avoid
versioning issues, since the other package needs to use the old
version of tastypie.

Lastly, in the urlconf we add the urls for the new app above the urls
for the old app.  The URLs overlap and this ensures that the URLs for
the new app will take precedence.