Implemented actions in the editor
Send actions to the editor via editor_data.

Refactered the session code to work with actions instead of tasks,
etc.  Added a bunch of tests for session code and also
SubtitleStorage.  Moved EditorData component out of
amara.SubtitleEditor.subtitles.services.  This allows us to mock it
out while testing SubtitleStorage

Implemented actions for tasks teams.

Added an in_progress_text attribute to Actions.  This is what we show
in the freeze box while waiting for the API request.

Pulled in a new version of the API to allow for sending actions when
posting subtitles.

Added visual_class for the Publish action