Fix add language link (#1687)
The issue was that we want to use the subtitle workflow to control how
we display the add language link, but subtitle workflows were specific
to a language.

Changed get_workflow to only input a video, not a video/language code.
Added method to control the add language section of the video page.

This means that workflows are now per-video, not per-video/language
combination.  It reduces the flexibility somewhat since we can't mix
workflows for a video (for example a tasks workflow for english and a
collaboration workflow for french).  However I don't think that's such
a big deal.