Slight refactor for subtitle workflow
Moved the code that handles pipeline.add_subtitles() being called to
the SubtitleWorkflow object.  This allows them to return different
actions if needed.  Also added a way for actions to set the visibility
for newly created SubtitleVersions.

For the collab model we need to be able to do this because we want to
return an action that will create private subtitles. We should be
doing this for the tasks code as well, but we have other code that
currently handles that and I didn't want to mess with it.