New code for Team workflows
Added some code to make team workflows more extensible.  TeamWorkflows
implement SubitleWorkflows and also alter the pages in the team's
section like the dashboard and workflow settings page.

The plan from here is to:
  - Add extra TeamWorkflow subclasses to amara-enterprise.  This is
    what we will use for the collab-style teams
  - Add extra TeamWorkflow subclasses to our main repository.  I think
    we can create one for public teams, one for private teams, and
    maybe others.

Removed the is_moderated field from the team create form.  We weren't
using that value or displaying the widget to select it.