Fixed unittests
This was mostly just switching them to use factory functions instead
of relying on fixtures, which are now deleted.

Removed MessagesTest.test_moderated_notifies_only_when_published.
That test was pretty confusing and also testing way to many things at
once.  It seems better to just let it go for now and replace it with a
nicer test if we need to.