Fixed issue updating Video.title with tasks (#1463)
Added code to fix it and also a migration to fix legacy data.  I've
already run the SQL for this migration a couple times on production so
when we deploy it shouldn't need to fix too many rows.

Also fixed issue running video_changed_tasks() when completing tasks.
I'm not sure what happened with the old code, but it looks like an
editing error caused totally crazy code to be there.