Better way to query counts for teams list
Use subselects instead of annotate (which uses group by) to count
members.  This runs much faster.

Use a similar query to count videos, which avoids an extra query for
each team on the list.

Renamed member_count() to members_count().  I think it's nicer for the
name to have the same form as videos_count().