Couple of fixes for setting titles
Set subtitle_visibility='private' for TaskAction.  This makes it so we
don't set the title before the subtitles are published.

Fixed how we set video metadata.  Before both the SubtitleLanguage and
the SubtitleVersion were trying to make changes to the video and save
it.  However, they both had a different video object, so the changes
were clobbering each other.