Added code to handle user cache
The main reason to do this is to optimize our memcache fetches.  Since
we know we are going to want to fetch several keys, we can use
get_many() to fetch all cache values at with one trip instead of
multiple ones.  It's probably not that important considering we only
have 2 keys right now, but it will get more important if we want to
cache other things per-user (which I hope we can do).

Also it simplifies the interface a bit and handles creating unique
keys per-user.