Removed a couple unneeded calls to can_user_see()
I'm going to refactor this code, so I figure we should delete the
cruft first.

We don't need it inside the video or language page, since the user
shouldn't be able te view the page at all if that's the case.

Deleted can_view_version() which is not used anymore.

Removed can_see_video() check from can_view_private_version(), since
the check for membership is a stronger assertion anyways.