Replaced Video.can_user_see() with subtitle workflow method
This gives us more flexibility, since we can define different checks
for each workflow.

Added a Video.get_workflow() method that gets a workflow and caches
the result.  Refactored code in the video view decorators to use the
new system.  Removed some old/dead code while I was at it.

There was one reference to can_user_see() in the teams permissions
code.  I'd rather not touch that code too much, instead we should just
delete most of it when we get rid of tasks.  Replaced the method call
with code that does the same thing.