Couple more message cacheing fixes
Delete the top-panel cache when we get a new message.  We want to
update the new message count in that case.

Don't try to cache User.unread_messages_count().  That code wasn't
correct because it will return the same value regardless of what
hidden_message_id is passed in.  In any case, the new caching code is
better so we can just get rid of this.