Couple of stastics fixes
Removed gauge_statistic_languages().  It was running slowly sometimes
and I don't think the data was needed.  The task was even using the
old SubtitleLanguage model, which indicates it hasn't been touched in
a while.

Fixed gauge_statistic() to not throw an error.  It now doesn't try to
log the view counts.