Make save draft an action
I think this simplifies things by making all things you can do in the
editor an action.  We have several workflows that want to perform some
work when a draft is saved and they have to override
action_for_add_subtitles(), which is annoying.  Now they can just add
the code to the save draft action.

We still need action_for_add_subtitles(), since we have to support API
calls from partners that don't specify an action.  But we now only
need to support it for older workflows.