Groundwork for better video page caching
Added the CacheGroup class, which will handle storing cached data for
a video.  I also plan to use it to replace the UserCache class.

Added a bunch of documentation about how it works.

Added a utility class to create codes like our video ids and editor
nonces.  I figure we use this often enough that it's good to put in
utils.