Speed up the add video to team section
It was actually fairly complex to calculate which teams you can move
the video to, so I moved it the form to a separate page.  The new page
doesn't look great but the old UI was to just dump you in the add
video form so I think it's not worse.

Store the result of the managed_teams() check in the user cache.

Moved the code to invalidate the user cache on team member changes to
the teams app.  I think it makes more sense to have it here.  Also
invalidate when the membership narrowings change.

Show the sharing buttons regardless if we are running from dev or not.
I'm not sure why we weren't but this makes things look more correct
while I'm testing the caching changes.

Added command to clear the cache