Add more functionality to CacheGroup
Added a bunch of code to integrate CacheGroup code with django model
class.  This should be able to replace the code in Video and
CustomUser to create a cache group for an instance.  In fact it should
be a nice improvement because with a cache pattern, we should be able
to fetch the instance and all cached values for that instance in using
a single get_many() call.

I think that this can achieve the dream of 0 DB queries for the video
page.  Using this code I think we should be able to handle the entire
page with the following 3 cache queries:

  - fetch the user and all its cached values from the cache
  - fetch the video and all it's cached values from the cache
  - (optionally) fetch the team and all its cached values from the
    cache.

Also added a convenience function to get a value from the cache and if
there's a cache miss compute it.