Speed up cache fetches
Removed the code to get the language choices from cache.  The code is
dirt simple so it's probably faster to calculate that fetch from
cache.  I made it so it's still cached in local memory though.

Refactored CustomUser.get_languages to return language codes instead
of UserLanguage objects.  That's how we're using it anyways.  Also
make it use the new caching code so that it can be pulled in with
get_many()

Use video.get_workflow() to avoid fetching the workflow twice.

Updated amara-enterprise revision to pull in a fix there.