More cache fixes
Fix call to request.use_cached_user() it was happening in the wrong
view decorator.

Cache comments and invalidate the video cache when a comment is saved.

Don't cache the video page headers.  This breaks when the user
switches tabs.  Instead cache the video url and comment counts.

Use video cache when calculating languages for the upload subtitles
form

Removed the SaveUserIP and ResponseTime middleware.  I don't think we
use this data at all, it's potentially slow, and it's messing with the
caching code.

Bumping version to pull in a fix from amara-enterprise