Use standard update_wrapper() call and remove redis code
This started as a project to use functools.update_wrapper instead of
the deprecated django version.  However, one of the modules that was
using it was redis_utils.  Since we don't actually have a redis server
anymore, I figured this was a good time to get rid of the old code
that depended on it.