Switch to new url template tag syntax
This is going to be used for the newer django versions, so let's
switch now.

The main change is quoting the viewname.  The other change is no
commas between the kwargs.

Added management command to check templates for warnings and compile
errors.

Removed profiles/actions_list.html.  I don't think we were using it
anymore and it had syntax errors.