Speedup the team activity page when filters are used
We were having an issue when the video language filter was used.  I'm
not really sure why, but it was changing how the mysql optimizer
handled the query and resulted in a filesort in a big temporary table.

Changed it so that we handle those filters as part of our subquery.
Refactored that code a bit and put it in the Team object instead of
ActionManager.  The main reason is that we now want to use TeamVideo
and we can't do that in videos.models because it's a circular import.