Log youtube get_video_info calls
We seem to be doing this too much.  Let's log the calls to see when
we're doing it.

Updated dev_settings.LOGGING.  There's really no reason for it to
redefine the variable, since the values from settings are pretty good.
The one thing we want to do in dev_settings is to disable the sentry
logging, since that's not available on dev setups.

Committing this directly to staging since it's a small change and I
believe that sylvain is out for this week.