Retry failed youtube sync attempts (#1858)
Added retry column to SyncHistory.  Added a task that periodically
picks on of the entries with retry=True and re-runs the sync for it.

Allow update_subtitles() to raise RetryableSyncingError.  If it does,
we will set retry=True.  Added code to do this for the youtube quota
errors.

Made the SyncHistory admin page nicer.