Test fixes after the changes
Fixed import in test_feeds.py

Fixed the test_delete_translation_tasks test to work with the new
action code.  This meant updating the test itself and also
TaskFactory.create_review().  Not sure why we didn't notice this one
before, it seems that the test wasn't running.

Removed the test_unpublish test.  I'm not really sure what it was
trying to do and it was definitely based off of old code.  Since it
didn't seem to be running either, I figure it's safe to just remove
it.