Fix the more link on the revisions tab
This was working by extending the code in
videos/templatetags/paginator.py, but then it stopped working after
some of my django 1.7 changes.

I think the paginator code is confusing and would love to get away
from it so I made it so the revisions tab doesn't use it and
implements the more link itself.