New deployment script
Created a script to handle deployment and cleanup.  We used to have
this in the config of several jenkins commands, but I think it's
easier to understand when it's in a script.

This script works basically like the old code, except it's in python
rather than shell scripts.  There are some changes though:

  - The deployment code can run migrations as part of the process.
  - There's an option to take down the app servers while migrations
    are running.
  - The script handles both deployment and cleanup.
  - The cleanup code is a bit different than the old cleanup command,
    mainly because it only tries to clean up containers that were
    started by the deploy script.
  - Removed the warm stage.  I don't think that it was needed and I
    remember there being issues with it once.

I've tested this a bunch with staging and it seems to work there.
Still need to test with production and preview branches.