Speedup for the video resource API
Added function to find which languages have a public version for using
without having to do a query per language.

This reduces the number of queries from ~70 to ~12 on my dev instance.
It could reduce them even more for some cases, since the number of
queries scaled with the number of languages for the video list.