Speedups for the subtitle language API endpoint
Did a bunch of work to eliminatate queries we need to do per-language
or per-version which is even worse.

Added method on the subtitle version manager to select all versions
for a list of languages and perform some extra optimizations.  Use
that at the begining of serialization to fetch all versions at once.

Added a test for fetch_for_languages() and updated the test for the
fetch_and_join() method using the new technique.

Added an optimization on SubtitleVersion to avoid a DB lookup if we
use prefetch_related('metadata') when fetching the versions.