Started new google login code
Took the youtube-specific code out of the oauth code in
externalsites.google.

Made the oauth login also send openid.realm and return the openid_id.
This is the old ID that we used to link accounts.  It also returns
sub, which will be the new ID used to link accounts.

The last step is to implement creating users and logging them in at
the end of the process.