Update for git commit parsing
Don't try to include the branch in commit.py.  We didn't really want
to use it anyways.

Use git rev-parse --short to generate short commit links.  This works
better than just slicing the list, since git can tell if the short
link is actually unique or not and add length to the id if needed.