Started implementing new stylesheet for new-style teams
This is an attempt to reboot our JS and CSS.  The new code is going to
live in media/src/css/site/ and media/src/js/site/.  It somewhat
matches our current styles, but not 100%.  Switched to jquery 2.x for
the new JS.

Made it so that when a non-member goes to the hompage for a new-style
team, they see a welcome page.  Right now that page just has some
placeholder text, but the header and footer are working.

Gave the homepage a name.