Added signals that fires before/after tests
This allows other components to connect to the signals and to reset
things.  This is useful for amara-enterprise stuff.  For the unisubs
repo we can just reset things in the plugin directly.

This was originally for amara-enterprise#122, but I actually don't
need it there so I'm just merging to staging.