Made the messages settings page use the new team views.
Renamed the view to "messages" since that's what we're calling it in
the UI.

Moved label text to the form code.  Use the _field.html template to
display inputs.

Added a decorator that renders either the old or new view depending on
the team type.

Fixed urlconf.  Before the new views were on top, but we want some old
views to take precedence.  For example we want to use the create view
for "/teams/create/", not try to render the dashboard for the "create"
team.