Implemented the project settings page using the new style
Added a couple more CSS styles for delete buttons and tables.

Moved some of the project logic away from the view and into the form
class.

Update for the project model code.  Now we update the slug when we
change the project name.  I'm not sure if this is right or not, but
the old way of keeping the old/now incorrect slug definitely seems
wrong.  Maybe we should have a slug form field so the user can choose
if they want to update or not.