Style/color cleanup
This is an attempt to move away from the hands-home colors and try
some new ones.

Made the menus/header 1 solid color.  This isn't supposed to be the
main point of focus for the page, so I figure having something that
doesn't jump out so much is good.

Changed button color (and actually there are several options to choose
from by changing which one is last in colors.css)

Made more things align at the left of the page.

Added a small margin between the home button and other menus.

If you open a menu, make the parent item highlighted.