More style changes
Make the user menu right-aligned so it doesn't go off the screen
(#2200)

Remove the metrics from the teams banner.  We don't need to show these
numbers for every page view, maybe we can have them just on the
dashboard.

Made nav menu items shorter.

Set the background for form inputs.  This makes it lighter for me on
chromium.