Simplify daily motion video URL matching (#2228)
Made it so DailymotionVideoType.matches_video_url() just checks the
form of the video URL, it doesn't actually try to make an API request.

This seems good in general to avoid requests and it also avoids the
exception in the API code.  The exception might happen in other
places, but I think that could be correct in those situations.