Changed how the team activity language filter works (#1930)
Made it match the new_language field on the activity table.  This
makes it work like you would expect with the add-comment, add-version,
and some other activity types.

One thing it doesn't let you do anymore is search for add-video
activity, and filter out videos that don't have subtitles for a given
language.  But I don't think that was really used anyhow.