Rename/Refactor the team members pages
Changed the name to "members" from "details_members".  It seems like
simpler is better to me.

Created a new url "team:admin-list" to view the list of admins.  This
page is for giving users a way to join invite-only teams.

Reworked the urlconf so that these pages go to the new_views module.
Right now we just forward the request right back to the old views.py
code.  The next step is implementing a new view for new-style teams.