Couple of DB improvements
This avoids an extra query when using request.user as django model
filter param.  I'm not exactly sure why we were creating the extra
query, but I think it has to do with how we implement User/CustomUser
and passing the "wrong" class to the query.  In any case, the User and
CustomUser instances share the same ID, so we can just use that.