Workaround for entrypoints bug in Docker 1.7.0 (Docker issue #14170)
Entrypoints that aren't expressed as lists are broken in Docker 1.7.0,
which causes the whole setup procedure to fail when using that
version. Expressing the entrypoints as list should be harmless, while
working around the issue at least until 1.7.1 is released. Tested
with Docker 1.7.0 and docker-compose 1.3.1.

More about the underlying bug:
https://github.com/docker/compose/issues/1594
https://github.com/docker/docker/issues/14170