Fixes for the permissions change on the videos page
The last commit made the team videos page members-only.  This wasn't
totally intentional, but I think it makes sense to do.  So we just
have to adjust the rest of the code to work with it.

Made it so that we link to the dashboard page instead of the team
videos page for the invite message and similar places.  On the welcome
message the dashboard link replaces both the videos and the tasks
link.

Fixed the unittests to login before checking that page.