Generalized the form setting logic of the members page
Made it so that we can set form values using html data attributes.  I
plan to use a very similar technique for the videos page, and I don't
want to have to write a function for each form.

Removed a bunch of commented out code that I forgot to delete before.