Added video feed settings page
I was going to add this to the videos page, but I think it makes more
sense as a settings page.

Added method to fill in some modal text when opening the dialog.

Removed username field from the feed form.  We weren't using that.  In
fact, we can't create feeds for youtube users anymore after the API v3
update.