Use modal dialogs on the project pages
This change makes it so we use modal dialogs for add/edit/delete
instead of sending the user to a different page.  I like the way this
was working for the video pages, so let's use it for projects as well.

Added way to make modal dialogs pop open at startup.  This makes it so
that if the form has an error, we can pop it back up.  Also, added a
way to clear the errors from a modal form.  This way if the user
closes the edit dialog for 1 project, then opens it for another, the
error list doesn't persist.