Reworked videos page
Made it so that we have a thumbnail that takes up the full height,
with lines of text to the right.  This gives us more space and allows
us to show the descriptions, which I believe was important to TED.  I
think this layout will also be nice for other pages including the
collab admin page and I think it will support bulk operations nicely.

Made the edit button more prominent and the move/remove buttons less
prominent.

Reworked paddings for h6 and paragraphs so that things line up evenly
with line height.  Having half a line height as padding means that we
may cut things off halfway into the line.

Reworked buttons to remove the border except when we need them.  This
allows us to calculate their size better since the height is based on
rems, but the border widths were in px.

Added styles to cut text off after n lines.
Added hover colors for buttons.
Made it so buttons can open modal dialogs too.
Made headers with links use dark gray instead of blue.