Added bulk move to the regular videos page
I think this is an improvement for a few reasons:

- The videos page and the bulk move videos page looked almost
  identical, which could cause confusion.  Users could think they were
  on one page, but were on another.  Also, they might not understand
  what the difference is at all.
- For power users, this eliminates a click.
- We could add other bulk actions pretty easily.  For example we could
  add a button to bulk set the primary audio language.

Added some general code to handle bulk-mode selection on thumb lists.
I hope to reuse a lot of this for the collaborations list.

Fixed TeamVideo.move_to() to trigger re-indexing when moving a video
to a different project in its current team.

Removed the new-style move_videos page.  It all happens on the regular
videos page now.

Slight update on where we put the modal close buttons in the DOM tree.

Couple of fixes for the MoveTeamVideos form code.

Conflicts:
	media/src/css/site/lists.scss