Use px instead of rem to size fonts.
The idea of rems is that the scale based on the users font size.  But
they can get pretty annoying to use and also 99% of the time when a
user scales a website, they do it with the zoom function  which will
scale both px and rem values.

I'm finally breaking down and just using px.  It's way simpler to
size things this way.