Last pass at the videos page
Changed it so that we have these forms on the videos page:
  - single video edit
  - bulk edit
  - bulk move
  - bulk remove

The bulk forms appear in a fixed position overlay at the bottom of the
screen.

I like this change for a few reasons:
  - Bulk edit and bulk remove seem like nice features
  - Once you select a video, the overlay is more obvious than adding a
    button in the sidebar.
  - Before we had duplicated functionality between the move video form
    and the bulk move videos form.  To make this worse, you could
    specify project in the bulk move video form, but not the non-bulk
    one.
  - Only displaying the edit form in the table means less visual
    noise.
  - The way the forms are separated feels nice to me.  Single
    video edit is for normal users.  The checkboxes/bulk edit forms
    are for admins.

Added a base class for the bulk forms.
Added method to copy in form inputs from the page when popping up a
modal form.

I've spent quite a while on this page, and it's not that important for
TED phase III, but I hope to use the same techniques for the collab
admin page and the dashboard page.