nav menu CSS refactor
As long as I'm refactoring, I figured I should fix an annoying thing:
having to add !important to css colors.  Refactored the header/menu
CSS to make it so we don't need it anymore.  I also removed it from
the button CSS, it seems we didn't need it there at all.