More small changes
Reworked openModal to input the data it uses to setup the modal.  This
allows us to pass in if we call that function directly.

Upgraded to fontawesome 4.4

Changed the video duration overlay colors slightly.

Removed the test for the move video form.  We now only have the bulk
move team videos form.

Moved the bottom sheet to the bottom of the HTML instead of inside the
content div.