Implemented simple team workflow
This can be used as a replacement for the Old workflow for non-tasks
teams.  The workflow is simple: team members can edit subs, non
members can't.  Non members can view videos/subtitles based on the
"videos public?" setting.

Changed the wording for the settings page from "Workflows" to
"Workflow".  There's only 1 workflow for a team, so I think this is
clearer.