Split up workflow code to allow pro-request integration
The reason for this change is that the current system doesn't allow
for easy integration of pro-requests.  We would need to add code to
each Workflow class to check if there's a pro-request and, if so,
modify the behavior of several methods.

The main issue is there's only 1 hook to use: get_workflow(), which
operates on a per-video bases.  But for pro-requests, we only want to
modify the behavior on a per-language basis.

The solution is to split up Workflow into 2 parts VideoWorkflow and
LanguageWorkflow.  Then if you want to modify things on a per-language
basis you can override get_language_workflow() instead of
get_workflow().

We already have a bunch of code that subclasses Workflow, and I don't
really have time to change it all now.  So instead, I added some
wrapper methods to VideoWorkflow so that it supports the same methods as
Workflow.  This means that we can have the old code return Workflow
and new code return VideoWorkflow and things will be compatible.  Only
code that returns a VideoWorkflow will be compatible with pro-requests
and get_language_workflow() in general, but I think that's okay for
now.

Right now, the only places where we return a VideoWorkflow subclasses
are non-team videos and the simple team workflow.  We should update
other code to use that as well.