Added team language page
This is very similar to the project page.  I tried to avoid repeating
code, but it was simpler to copy in some places.

There is no way to navigate here for normal teams.  This should be
part of the simple teams dashboard when we implement it.  The main
point is to allow collab teams to use it.