Tooltip improvements
Allow different tooltips based on if the container overflows or not
This is something that we're using a lot in the collab UI.

Positioned the tooltips a bit differently so that it's possible to
move the pointer over the tooltip without it disappearing.

Don't let tooltips go too far to the right of the page.

Fixed tooltips inside elements with the one-line class.