Dropdown improvements
Make dropdown button menus expand to their parent's width
Close the dropdown when a link inside it is clicked.
Make all links change the cursor on hover.  This is especially nice
for dropdown links that open modals.
Keep event handlers around when moving the dropdown to the toplevel.
Fix for corner case with autocomplete lists