Moved multi-form handling to the utils package
We are using this technique in a few places so I figured it makes
sense to move it out of the view function and into a separate module.
It also adds some features that I want to use like lazy form creation.
And lastly, it simplifies the view code.

Made all the forms on the videos page input the videos_qs in their
constructor.  This allows us to simplfy things by having the same
method signatures for all forms.