Refactored code to enable/disable forms on the videos page
Moved the code out of the forms and into a separate function.  The
main reason to do this is to simplfy the form code.  This is useful
becuase I want to make it so other apps can insert forms in the team
videos page.  It also makes things easier to test.