Use AJAX for forms on the videos page
Switched the videos page from renderirg forms on the client-side with a
combination of HTML and JS.  Now the forms are rendered server side
and we use AJAX to fetch/submit them.

The main reason for this is that it makes it the system more
easily extensible.  Another app can add a form to the videos page
without us needing to worry about making JS code to handle things.

A side benefit is that we can use django template logic, which makes
the display a bunch nicer, especially when handling pluralization.

This change means we don't need the FormRouter class that I wrote in
the last commit, but I still think it's useful to keep around.