Couple changes/fixes
Fixed some CSS points.

Started working on a field to handle multiple language selection, but
didn't finish it.  I would love to make this work with something like
sumoselect, but the way that's coded doesn't work with our modal
dialogs.

In the end, I just made it work with a regular SelectMultiple widget,
but having the field there is good because it paves the way to change
the input in the future.